% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fittingFunction.R
\name{evaluateCompetitionModel}
\alias{evaluateCompetitionModel}
\title{evaluateCompetitionModel}
\usage{
evaluateCompetitionModel(dataSet = cpcRawData, model, fitFunction = F)
}
\arguments{
\item{dataSet}{A tibble with whatever headers are needed to fit the model, but which has to at least include}

\item{model}{A model that takes whatever dataSet is, and creates a new column, by default named "modelProjectionB",}

\item{fitFunction}{A function that integrates the outcome of the model application and the dataSet. Defaults to mean squared error}
}
\value{
A list containing a tibble and a fit outcome.
}
\description{
This simple script just applies a model entry (which matches a particular form) to a data set, using an evaluation measure. All of these can be specified, but have default functions
}
